内部処理調整
