+++ php/std_output.php
@@ -1,0 +1,226 @@
+<?php
+/**
+ * px2-px2dthelper
+ */
+namespace tomk79\pickles2\px2dthelper;
+
+/**
+ * std_output.php
+ */
+class std_output{
+
+	/** Picklesオブジェクト */
+	private $px;
+
+	/**
+	 * constructor
+	 *
+	 * @param object $px $pxオブジェクト
+	 */
+	public function __construct( $px ){
+		$this->px = $px;
+	}
+
+	/**
+	 * データを自動的に加工して返す。
+	 *
+	 * @param mixed $val 加工するデータ
+	 * @return string 加工されたテキストデータ
+	 */
+	public function data_convert($val){
+		$data_type = $this->px->req()->get_param('type');
+		if( !is_string($data_type) || !strlen($data_type) ){
+			$data_type = 'json';
+		}
+		if( $data_type == 'json' ){
+			header('Content-type: application/json; charset=UTF-8');
+		}elseif( $data_type == 'jsonp' ){
+			header('Content-type: application/javascript; charset=UTF-8');
+		}elseif( $data_type == 'xml' ){
+			header('Content-type: application/xml; charset=UTF-8');
+		}
+		switch( $data_type ){
+			case 'jsonp':
+				return $this->data2jsonp($val);
+				break;
+			case 'json':
+				return $this->data2json($val);
+				break;
+			case 'xml':
+				return $this->data2xml($val);
+				break;
+		}
+		// return self::data2jssrc($val);
+		return $val;
+	}
+
+	/**
+	 * データをXMLに加工して返す。
+	 *
+	 * @param mixed $val 加工するデータ
+	 * @return string 加工されたテキストデータ
+	 */
+	private function data2xml($val){
+		return '<api>'.self::xml_encode($val).'</api>';
+	}
+
+	/**
+	 * データをJSONに加工して返す。
+	 *
+	 * @param mixed $val 加工するデータ
+	 * @return string 加工されたテキストデータ
+	 */
+	private function data2json($val){
+		// return self::data2jssrc($val);
+		return json_encode($val);
+	}
+
+	/**
+	 * データをJSONPに加工して返す。
+	 *
+	 * @param mixed $val 加工するデータ
+	 * @return string 加工されたテキストデータ
+	 */
+	private function data2jsonp($val){
+		//JSONPのコールバック関数名は、パラメータ callback に受け取る。
+		$cb = trim( $this->px->req()->get_param('callback') );
+		if( !strlen($cb) ){
+			$cb = 'callback';
+		}
+		// return $cb.'('.self::data2jssrc($val).');';
+		return $cb.'('.json_encode($val).');';
+	}
+
+
+	/**
+	 * 変数をXML構造に変換する
+	 *
+	 * @param mixed $value 値
+	 * @param array $options オプション
+	 * <dl>
+	 *   <dt>delete_arrayelm_if_null</dt>
+	 *     <dd>配列の要素が `null` だった場合に削除。</dd>
+	 *   <dt>array_break</dt>
+	 *     <dd>配列に適当なところで改行を入れる。</dd>
+	 * </dl>
+	 * @return string XMLシンタックスに変換された値
+	 */
+	private static function xml_encode( $value = null , $options = array() ){
+
+		if( is_array( $value ) ){
+			#	配列
+			$is_hash = false;
+			$i = 0;
+			foreach( $value as $key=>$val ){
+				#	ArrayかHashか見極める
+				if( !is_int( $key ) ){
+					$is_hash = true;
+					break;
+				}
+				if( $key != $i ){
+					#	順番通りに並んでなかったらHash とする。
+					$is_hash = true;
+					break;
+				}
+				$i ++;
+			}
+
+			if( $is_hash ){
+				$RTN .= '<object>';
+			}else{
+				$RTN .= '<array>';
+			}
+			if( $options['array_break'] ){ $RTN .= "\n"; }
+			foreach( $value as $key=>$val ){
+				if( $options['delete_arrayelm_if_null'] && is_null( $value[$key] ) ){
+					#	配列のnull要素を削除するオプションが有効だった場合
+					continue;
+				}
+				$RTN .= '<element';
+				if( $is_hash ){
+					$RTN .= ' name="'.htmlspecialchars( $key ).'"';
+				}
+				$RTN .= '>';
+				$RTN .= self::xml_encode( $value[$key] , $options );
+				$RTN .= '</element>';
+				if( $options['array_break'] ){ $RTN .= "\n"; }
+			}
+			if( $is_hash ){
+				$RTN .= '</object>';
+			}else{
+				$RTN .= '</array>';
+			}
+			if( $options['array_break'] ){ $RTN .= "\n"; }
+			return	$RTN;
+		}
+
+		if( is_object( $value ) ){
+			#	オブジェクト型
+			$RTN = '';
+			$RTN .= '<object>';
+			$proparray = get_object_vars( $value );
+			$methodarray = get_class_methods( get_class( $value ) );
+			foreach( $proparray as $key=>$val ){
+				$RTN .= '<element name="'.htmlspecialchars( $key ).'">';
+
+				$RTN .= self::xml_encode( $val , $options );
+				$RTN .= '</element>';
+			}
+			$RTN .= '</object>';
+			return	$RTN;
+		}
+
+		if( is_int( $value ) ){
+			#	数値
+			$RTN = '<value type="int">'.htmlspecialchars( $value ).'</value>';
+			return	$RTN;
+		}
+
+		if( is_float( $value ) ){
+			#	浮動小数点
+			$RTN = '<value type="float">'.htmlspecialchars( $value ).'</value>';
+			return	$RTN;
+		}
+
+		if( is_string( $value ) ){
+			#	文字列型
+			$RTN = '<value type="string">'.htmlspecialchars( $value ).'</value>';
+			return	$RTN;
+		}
+
+		if( is_null( $value ) ){
+			#	ヌル
+			return	'<value type="null"></value>';
+		}
+
+		if( is_resource( $value ) ){
+			#	リソース型
+			return	'<value type="undefined"></value>';
+		}
+
+		if( is_bool( $value ) ){
+			#	ブール型
+			if( $value ){
+				return	'<value type="bool">true</value>';
+			}else{
+				return	'<value type="bool">false</value>';
+			}
+		}
+
+		return	'<value type="undefined"></value>';
+
+	}
+
+	/**
+	 * ダブルクオートで囲えるようにエスケープ処理する。
+	 *
+	 * @param string $text テキスト
+	 * @return string エスケープされたテキスト
+	 */
+	private static function escape_doublequote( $text ){
+		$text = preg_replace( '/\\\\/' , '\\\\\\\\' , $text);
+		$text = preg_replace( '/"/' , '\\"' , $text);
+		return	$text;
+	}
+
+}
